Separate out object metadata
